externalize GudhiRandomGenerator to be used by some other modules
