+++ src/python/gudhi/random.py
@@ -12,7 +12,7 @@ __license__ = "LGPL v3"
 
 from typing import Optional
 
-from gudhi import _random_ext as t
+from gudhi._random_ext import GudhiRandomGenerator
 
 import numpy as np
 
@@ -20,7 +20,7 @@ class GudhiBitGenerator(np.random.BitGenerator):
     def __init__(self, seed: Optional[int] = None):
         super().__init__(0)
         if seed is None:
-            self.rng = t.RandomGenerator()
+            self.rng = GudhiRandomGenerator()
         else:
-            self.rng = t.RandomGenerator(seed)
+            self.rng = GudhiRandomGenerator(seed)
         self.rng.setup_bitgen(self.capsule)
